The project structure was changed: - deleted files:    - API-methods.js    - Axios.js    // all the methods from these files were added to new file apiMethods    - deleted file methods.js - all the methods from this file were added to a new file uiMethods  - in the files login.spec.js and profileTest.spec.js the paths of inheritance were changed  - added new files to the directory 'helpers':     - apiMethods     - uiMethods
All tests passed after these changes (see screenshot)
